Proof of concept now runs as user on mac.
The auto-start system on mac is weird. it doesn't run as root. it runs as the user that installs it. If installed as root, it would only run if the root user logs in. This is a pain, but not the end of the world. Does break some functionality  (like being able to remote shutdown).